+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_ar.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=الفضاء:
 notes.label.movePageCurrentPosition=الموقع الحالي
 notes.label.movePageDestination=الوجهة:
 notes.label.movePagePosition=المهنة:
-notes.label.noteTreeview.tooltip=انقر لعرض شجرة المعلومات
 notes.label.historyVersions=الإصدارات
 notes.label.restoreVersions=استعادة
 notes.untitled.title=بدون عنوان
