+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_ca.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Espai:
 notes.label.movePageCurrentPosition=Posició actual:
 notes.label.movePageDestination=Destí:
 notes.label.movePagePosition=Posició:
-notes.label.noteTreeview.tooltip=Feu clic per mostrar l'arbre de notes
 notes.label.historyVersions=Versions
 notes.label.restoreVersions=Restaura
 notes.untitled.title=Sense títol
