+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_co.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Spaziu:
 notes.label.movePageCurrentPosition=Posizione attuale:
 notes.label.movePageDestination=Destination:
 notes.label.movePagePosition=Posizione:
-notes.label.noteTreeview.tooltip=Cliccate per vede l'arburu Notes
 notes.label.historyVersions=Versioni
 notes.label.restoreVersions=Risturà
 notes.untitled.title=Sans titre
