+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_cs.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Skupina:
 notes.label.movePageCurrentPosition=Aktuální pozice:
 notes.label.movePageDestination=Cíl:
 notes.label.movePagePosition=Pozice:
-notes.label.noteTreeview.tooltip=Kliknutím zobrazíte strom poznámek
 notes.label.historyVersions=Verze
 notes.label.restoreVersions=Obnovit
 notes.untitled.title=Bez názvu
