+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_de.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Projektraum
 notes.label.movePageCurrentPosition=Aktuelle Position:
 notes.label.movePageDestination=Ziel:
 notes.label.movePagePosition=Position:
-notes.label.noteTreeview.tooltip=Klicken, um Notizbaum anzuzeigen
 notes.label.historyVersions=Versionen
 notes.label.restoreVersions=Wiederherstellen
 notes.untitled.title=Ohne Titel
