+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_el.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Χώρος:
 notes.label.movePageCurrentPosition=Τρέχουσα θέση:
 notes.label.movePageDestination=Προορισμός:
 notes.label.movePagePosition=Θέση:
-notes.label.noteTreeview.tooltip=Κλικ για εμφάνιση δέντρου σημειώσεων
 notes.label.historyVersions=Εκδόσεις
 notes.label.restoreVersions=Επαναφορά
 notes.untitled.title=Χωρίς τίτλο
