+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_en.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Space:
 notes.label.movePageCurrentPosition=Current position:
 notes.label.movePageDestination=Destination:
 notes.label.movePagePosition=Position:
-notes.label.noteTreeview.tooltip=Click to display Notes tree
 notes.label.historyVersions=Versions
 notes.label.restoreVersions=Restore
 notes.untitled.title=Untitled
@@ -175,6 +174,11 @@ notes.label.no-content.no-redactor.content.last=Or create a new note
 notes.label.noteNotFound=We could not find this note.
 notes.label.noteSearchNotFound=No notes found for:
 notes.label.noteNotFound.button=Back to Notes App
+
+notes.label.tree=Treeview
+notes.tooltip.close.tree=Hide treeview
+notes.tooltip.open.tree=Display treeview
+
 notes.message.missingImportNotes=Import Notes is mandatory.
 import.drawer.maxFileSize.error=You can’t upload files larger than {0} MB
 import.drawer.fileType.error=You should upload a zip note
