+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_es_ES.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Espacio:
 notes.label.movePageCurrentPosition=Posición actual:
 notes.label.movePageDestination=Destino:
 notes.label.movePagePosition=Posición:
-notes.label.noteTreeview.tooltip=Haz clic para mostrar el árbol de notas
 notes.label.historyVersions=Versiones
 notes.label.restoreVersions=Restaurar
 notes.untitled.title=Sin título
