+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_et.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Tühik:
 notes.label.movePageCurrentPosition=Praegune positsioon:
 notes.label.movePageDestination=Sihtkoht:
 notes.label.movePagePosition=Asukoht:
-notes.label.noteTreeview.tooltip=Klõpsake märkmete puu kuvamiseks
 notes.label.historyVersions=Versioonid
 notes.label.restoreVersions=Taasta
 notes.untitled.title=Pealkirjata
