+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_fa.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=فضا:
 notes.label.movePageCurrentPosition=موقعیت فعلی:
 notes.label.movePageDestination=مقصد:
 notes.label.movePagePosition=موقعیت:
-notes.label.noteTreeview.tooltip=برای نمایش درخت یادداشت ها کلیک کنید
 notes.label.historyVersions=نسخه ها
 notes.label.restoreVersions=بازيابي
 notes.untitled.title=بدون عنوان
