+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_fi.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Tila:
 notes.label.movePageCurrentPosition=Nykyinen sijainti:
 notes.label.movePageDestination=Kohde:
 notes.label.movePagePosition=Sijainti:
-notes.label.noteTreeview.tooltip=Klikkaa näyttääksesi muistiinpanojen puu
 notes.label.historyVersions=Versiot
 notes.label.restoreVersions=Palauta
 notes.untitled.title=Nimetön
