+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_fil.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Ang espasyo:
 notes.label.movePageCurrentPosition=Kasalukuyang posisyon:
 notes.label.movePageDestination=Ang destinasyon:
 notes.label.movePagePosition=Ang Posisyon:
-notes.label.noteTreeview.tooltip=I-click upang ipakita ang Notes tree
 notes.label.historyVersions=Ang mga bersyon
 notes.label.restoreVersions=Pagpapanumbalik
 notes.untitled.title=Walang Pamagat
