+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_fr.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Espace :
 notes.label.movePageCurrentPosition=Emplacement actuel : 
 notes.label.movePageDestination=Destination :
 notes.label.movePagePosition=Position :
-notes.label.noteTreeview.tooltip=Cliquer pour afficher l'arborescence des notes
 notes.label.historyVersions=Versions
 notes.label.restoreVersions=Restaurer
 notes.untitled.title=Sans titre
@@ -175,6 +174,11 @@ notes.label.no-content.no-redactor.content.last=Ou créer une nouvelle note
 notes.label.noteNotFound=Nous n'avons pas pu trouver cette note.
 notes.label.noteSearchNotFound=Aucune note trouvée pour :
 notes.label.noteNotFound.button=Revenir à l'appli Notes
+
+notes.label.tree=Arborescence
+notes.tooltip.close.tree=Masquer l'arborescence
+notes.tooltip.open.tree=Afficher treeview
+
 notes.message.missingImportNotes=Importer des notes est obligatoire.
 import.drawer.maxFileSize.error=Vous ne pouvez pas charger des fichiers de plus de {0} Mo
 import.drawer.fileType.error=Vous devez sélectionner un fichier au format zip
