+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_he.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=חלל:
 notes.label.movePageCurrentPosition=תפקיד נוכחי:
 notes.label.movePageDestination=יַעַד:
 notes.label.movePagePosition=מַצָב:
-notes.label.noteTreeview.tooltip=לחץ כדי להציג את עץ ההערות
 notes.label.historyVersions=גרסאות
 notes.label.restoreVersions=לְשַׁחְזֵר
 notes.untitled.title=ללא כותרת
