+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_hr.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Prostor:
 notes.label.movePageCurrentPosition=Trenutačna pozicija:
 notes.label.movePageDestination=Odredište:
 notes.label.movePagePosition=Položaj:
-notes.label.noteTreeview.tooltip=Kliknite za prikaz stabla Bilješki
 notes.label.historyVersions=Verzije
 notes.label.restoreVersions=Vratiti
 notes.untitled.title=Bez naslova
