+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_hu.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Űr:
 notes.label.movePageCurrentPosition=Jelenlegi pozíció:
 notes.label.movePageDestination=Rendeltetési hely:
 notes.label.movePagePosition=Pozíció:
-notes.label.noteTreeview.tooltip=Kattintson a jegyzetfa megjelenítéséhez
 notes.label.historyVersions=Verziók
 notes.label.restoreVersions=Visszaállítás
 notes.untitled.title=Cím nélkül
