+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_id.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Forum:
 notes.label.movePageCurrentPosition=Posisi saat ini:
 notes.label.movePageDestination=Tujuan:
 notes.label.movePagePosition=Posisi:
-notes.label.noteTreeview.tooltip=Klik untuk menampilkan pohon Catatan.
 notes.label.historyVersions=Versi
 notes.label.restoreVersions=Pulihan
 notes.untitled.title=Tanpa judul
