+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_in.properties
@@ -88,7 +88,6 @@ notes.label.movePageSpace=Space:
 notes.label.movePageCurrentPosition=Current position:
 notes.label.movePageDestination=Destination:
 notes.label.movePagePosition=Position:
-notes.label.noteTreeview.tooltip=Click to display Notes tree
 notes.label.historyVersions=Versions
 notes.label.restoreVersions=Restore
 notes.untitled.title=Untitled
