+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_it.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Spazio:
 notes.label.movePageCurrentPosition=Posizione corrente:
 notes.label.movePageDestination=Destinazione:
 notes.label.movePagePosition=Posizione:
-notes.label.noteTreeview.tooltip=Cliccare per visualizzare l'albero delle Note
 notes.label.historyVersions=Gestisci Versioni
 notes.label.restoreVersions=Ripristina
 notes.untitled.title=Senza Titolo
