+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_ja.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=スペース:
 notes.label.movePageCurrentPosition=現在の位置:
 notes.label.movePageDestination=保存先：
 notes.label.movePagePosition=ポジション:
-notes.label.noteTreeview.tooltip=クリックしてメモツリーを表示する
 notes.label.historyVersions=バージョン管理
 notes.label.restoreVersions=リストア
 notes.untitled.title=タイトルなし
