+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_ko.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=공간:
 notes.label.movePageCurrentPosition=현재 위치:
 notes.label.movePageDestination=목적지:
 notes.label.movePagePosition=위치:
-notes.label.noteTreeview.tooltip=클릭하여 메모 트리를 표시합니다.
 notes.label.historyVersions=버전
 notes.label.restoreVersions=복원하다
 notes.untitled.title=제목 없음
