+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_lt.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Erdvė:
 notes.label.movePageCurrentPosition=Dabartinė padėtis:
 notes.label.movePageDestination=Paskirtis:
 notes.label.movePagePosition=Pozicija:
-notes.label.noteTreeview.tooltip=Spustelėkite, kad būtų rodomas užrašų medis
 notes.label.historyVersions=Versijos
 notes.label.restoreVersions=Atkurti
 notes.untitled.title=Be pavadinimo
