+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_nl.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Ruimte:
 notes.label.movePageCurrentPosition=Huidige positie:
 notes.label.movePageDestination=Bestemming:
 notes.label.movePagePosition=Positie:
-notes.label.noteTreeview.tooltip=Klik om Notitieboom weer te geven
 notes.label.historyVersions=Versies beheren
 notes.label.restoreVersions=Herstellen
 notes.untitled.title=Naamloos
