+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_no.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Fellesskap:
 notes.label.movePageCurrentPosition=Nåværende posisjon:
 notes.label.movePageDestination=Destinasjon:
 notes.label.movePagePosition=Stilling:
-notes.label.noteTreeview.tooltip=Klikk for å vise Notattre
 notes.label.historyVersions=Versjoner
 notes.label.restoreVersions=Gjenopprett
 notes.untitled.title=Uten tittel
