+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_pl.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Przestrzeń:
 notes.label.movePageCurrentPosition=Bieżąca pozycja:
 notes.label.movePageDestination=Cel:
 notes.label.movePagePosition=Stanowisko:
-notes.label.noteTreeview.tooltip=Kliknij, aby wyświetlić drzewo notatek
 notes.label.historyVersions=Wersje
 notes.label.restoreVersions=Przywracanie
 notes.untitled.title=Bez tytułu
