+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_pt_PT.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Espaço:
 notes.label.movePageCurrentPosition=Posição atual:
 notes.label.movePageDestination=Destino:
 notes.label.movePagePosition=Cargo:
-notes.label.noteTreeview.tooltip=Clique para exibir a árvore de notas
 notes.label.historyVersions=Versões
 notes.label.restoreVersions=Restaurar
 notes.untitled.title=Sem título
