+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_ro.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Spaţiu:
 notes.label.movePageCurrentPosition=Poziţia curentă:
 notes.label.movePageDestination=Destinație:
 notes.label.movePagePosition=Poziţie:
-notes.label.noteTreeview.tooltip=Click pentru a afişa note arborescente
 notes.label.historyVersions=Versiuni
 notes.label.restoreVersions=Restabilire
 notes.untitled.title=Fără Nume
