+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_ru.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Область:
 notes.label.movePageCurrentPosition=Должность в настоящее время:
 notes.label.movePageDestination=Назначение:
 notes.label.movePagePosition=Должность:
-notes.label.noteTreeview.tooltip=Нажмите, чтобы отобразить дерево записей
 notes.label.historyVersions=Версии
 notes.label.restoreVersions=Восстановить
 notes.untitled.title=Без названия
