+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_sk.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Priestor:
 notes.label.movePageCurrentPosition=Aktuálna pozícia:
 notes.label.movePageDestination=Cieľ:
 notes.label.movePagePosition=Pozícia:
-notes.label.noteTreeview.tooltip=Kliknutím zobrazíte strom poznámok
 notes.label.historyVersions=Verzie
 notes.label.restoreVersions=Obnoviť
 notes.untitled.title=Bez názvu
