+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_sl.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Prostor:
 notes.label.movePageCurrentPosition=Trenutni položaj:
 notes.label.movePageDestination=Cilj:
 notes.label.movePagePosition=Delovno mesto:
-notes.label.noteTreeview.tooltip=Kliknite za prikaz drevesa Notes
 notes.label.historyVersions=Različice
 notes.label.restoreVersions=Obnovi
 notes.untitled.title=Neimenovan
