+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_sq.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=crwdns44762:0crwdne44762:0
 notes.label.movePageCurrentPosition=crwdns44764:0crwdne44764:0
 notes.label.movePageDestination=crwdns44766:0crwdne44766:0
 notes.label.movePagePosition=crwdns44768:0crwdne44768:0
-notes.label.noteTreeview.tooltip=crwdns44770:0crwdne44770:0
 notes.label.historyVersions=crwdns44772:0crwdne44772:0
 notes.label.restoreVersions=crwdns44774:0crwdne44774:0
 notes.untitled.title=crwdns44776:0crwdne44776:0
