+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_sv_SE.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Webb-yta:
 notes.label.movePageCurrentPosition=Nuvarande befattning:
 notes.label.movePageDestination=Destination:
 notes.label.movePagePosition=Befattning:
-notes.label.noteTreeview.tooltip=Klicka för att visa Anteckningsträd
 notes.label.historyVersions=Versioner
 notes.label.restoreVersions=Återställ
 notes.untitled.title=Namnlös
