+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_th.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=อวกาศ:
 notes.label.movePageCurrentPosition=ตำแหน่งปัจจุบัน :
 notes.label.movePageDestination=ปลายทาง:
 notes.label.movePagePosition=ตำแหน่ง:
-notes.label.noteTreeview.tooltip=คลิกเพื่อแสดงแผนผังบันทึก
 notes.label.historyVersions=เวอร์ชันต่างๆ
 notes.label.restoreVersions=คืนค่า
 notes.untitled.title=ไม่มีชื่อ
