+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_tr.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Alan:
 notes.label.movePageCurrentPosition=Şu anki pozisyonu:
 notes.label.movePageDestination=Hedef:
 notes.label.movePagePosition=Pozisyon:
-notes.label.noteTreeview.tooltip=Notlar ağacını görüntülemek için tıklayın
 notes.label.historyVersions=Sürümler
 notes.label.restoreVersions=Geri Yükle
 notes.untitled.title=Başlıksız
