+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_uk.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Область:
 notes.label.movePageCurrentPosition=Поточна позиція:
 notes.label.movePageDestination=Призначення:
 notes.label.movePagePosition=Посада:
-notes.label.noteTreeview.tooltip=Натисніть, щоб відобразити дерево нотаток
 notes.label.historyVersions=Версії
 notes.label.restoreVersions=Відновити
 notes.untitled.title=Без назви
