+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_vi.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=Không gian Nhóm:
 notes.label.movePageCurrentPosition=Vị trí hiện tại:
 notes.label.movePageDestination=Điểm đến:
 notes.label.movePagePosition=Vị trí:
-notes.label.noteTreeview.tooltip=Nhấp để hiển thị cây Ghi chú
 notes.label.historyVersions=Phiên bản
 notes.label.restoreVersions=Khôi phục
 notes.untitled.title=Không có Tiêu đề
