+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_zh_CN.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=空间:
 notes.label.movePageCurrentPosition=现任职务：
 notes.label.movePageDestination=目的地：
 notes.label.movePagePosition=位置：
-notes.label.noteTreeview.tooltip=点击显示笔记树
 notes.label.historyVersions=版本
 notes.label.restoreVersions=恢复
 notes.untitled.title=无标题
