+++ notes-webapp/src/main/resources/locale/portlet/notes/notesPortlet_zh_TW.properties
@@ -159,7 +159,6 @@ notes.label.movePageSpace=空間:
 notes.label.movePageCurrentPosition=目前位置：
 notes.label.movePageDestination=目的地：
 notes.label.movePagePosition=位置：
-notes.label.noteTreeview.tooltip=按一下可顯示註釋樹
 notes.label.historyVersions=版本
 notes.label.restoreVersions=還原
 notes.untitled.title=無標題
